+++ prince-archiver/prince_archiver/entrypoints/timestep_generator.py
@@ -7,7 +7,8 @@ from uuid import uuid4
 
 from prince_archiver.dto import TimestepMeta
 from prince_archiver.logging import configure_logging
-from prince_archiver.utils import make_timestep_directory, now
+from prince_archiver.test_utils import make_timestep_directory
+from prince_archiver.utils import now
 
 
 def _create_meta() -> TimestepMeta:
